+++ cardano-cli/cardano-cli.cabal
@@ -61,6 +61,19 @@ library
                         Cardano.CLI.Render
                         Cardano.CLI.Run
                         Cardano.CLI.Run.Legacy
+                        Cardano.CLI.Run.Legacy.Address
+                        Cardano.CLI.Run.Legacy.Address.Info
+                        Cardano.CLI.Run.Legacy.Genesis
+                        Cardano.CLI.Run.Legacy.Governance
+                        Cardano.CLI.Run.Legacy.Key
+                        Cardano.CLI.Run.Legacy.Node
+                        Cardano.CLI.Run.Legacy.Pool
+                        Cardano.CLI.Run.Legacy.Query
+                        Cardano.CLI.Run.Legacy.StakeAddress
+                        Cardano.CLI.Run.Legacy.TextView
+                        Cardano.CLI.Run.Legacy.Transaction
+                        Cardano.CLI.Run.Legacy.Read
+                        Cardano.CLI.Run.Legacy.Validate
                         Cardano.CLI.Types
 
                         Cardano.CLI.Byron.Commands
@@ -84,19 +97,6 @@ library
                         Cardano.CLI.Shelley.Orphans
                         Cardano.CLI.Shelley.Output
                         Cardano.CLI.Shelley.Parsers
-                        Cardano.CLI.Shelley.Run.Address
-                        Cardano.CLI.Shelley.Run.Address.Info
-                        Cardano.CLI.Shelley.Run.Genesis
-                        Cardano.CLI.Shelley.Run.Governance
-                        Cardano.CLI.Shelley.Run.Key
-                        Cardano.CLI.Shelley.Run.Node
-                        Cardano.CLI.Shelley.Run.Pool
-                        Cardano.CLI.Shelley.Run.Query
-                        Cardano.CLI.Shelley.Run.StakeAddress
-                        Cardano.CLI.Shelley.Run.TextView
-                        Cardano.CLI.Shelley.Run.Transaction
-                        Cardano.CLI.Shelley.Run.Read
-                        Cardano.CLI.Shelley.Run.Validate
 
                         Cardano.CLI.Ping
