+++ cardano-cli/src/Cardano/CLI/Run/Legacy.hs
@@ -9,17 +9,17 @@ module Cardano.CLI.Run.Legacy
 import           Cardano.Api
 
 import           Cardano.CLI.Conway.Commands
+import           Cardano.CLI.Run.Legacy.Address
+import           Cardano.CLI.Run.Legacy.Genesis
+import           Cardano.CLI.Run.Legacy.Governance
+import           Cardano.CLI.Run.Legacy.Key
+import           Cardano.CLI.Run.Legacy.Node
+import           Cardano.CLI.Run.Legacy.Pool
+import           Cardano.CLI.Run.Legacy.Query
+import           Cardano.CLI.Run.Legacy.StakeAddress
+import           Cardano.CLI.Run.Legacy.TextView
+import           Cardano.CLI.Run.Legacy.Transaction
 import           Cardano.CLI.Shelley.Parsers
-import           Cardano.CLI.Shelley.Run.Address
-import           Cardano.CLI.Shelley.Run.Genesis
-import           Cardano.CLI.Shelley.Run.Governance
-import           Cardano.CLI.Shelley.Run.Key
-import           Cardano.CLI.Shelley.Run.Node
-import           Cardano.CLI.Shelley.Run.Pool
-import           Cardano.CLI.Shelley.Run.Query
-import           Cardano.CLI.Shelley.Run.StakeAddress
-import           Cardano.CLI.Shelley.Run.TextView
-import           Cardano.CLI.Shelley.Run.Transaction
 
 import           Control.Monad.Trans.Except (ExceptT)
 import           Control.Monad.Trans.Except.Extra (firstExceptT)
