+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Node.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 
-module Cardano.CLI.Shelley.Run.Node
+module Cardano.CLI.Run.Legacy.Node
   ( ShelleyNodeCmdError(ShelleyNodeCmdReadFileError)
   , renderShelleyNodeCmdError
   , runNodeCmd
