+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Pool.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RankNTypes #-}
 
-module Cardano.CLI.Shelley.Run.Pool
+module Cardano.CLI.Run.Legacy.Pool
   ( ShelleyPoolCmdError(ShelleyPoolCmdReadFileError)
   , renderShelleyPoolCmdError
   , runPoolCmd
