+++ cardano-cli/src/Cardano/CLI/Run/Legacy/Transaction.hs
@@ -9,7 +9,7 @@
 
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
 
-module Cardano.CLI.Shelley.Run.Transaction
+module Cardano.CLI.Run.Legacy.Transaction
   ( ShelleyTxCmdError(..)
   , renderShelleyTxCmdError
   , runTransactionCmd
@@ -24,11 +24,11 @@ import           Cardano.Api.Shelley
 import           Cardano.CLI.Conway.Types
 import           Cardano.CLI.Helpers (printWarning)
 import           Cardano.CLI.Json.Friendly (friendlyTxBS, friendlyTxBodyBS)
+import           Cardano.CLI.Run.Legacy.Genesis
+import           Cardano.CLI.Run.Legacy.Read
+import           Cardano.CLI.Run.Legacy.Validate
 import           Cardano.CLI.Shelley.Output
 import           Cardano.CLI.Shelley.Parsers
-import           Cardano.CLI.Shelley.Run.Genesis
-import           Cardano.CLI.Shelley.Run.Read
-import           Cardano.CLI.Shelley.Run.Validate
 import           Cardano.CLI.Types
 import           Ouroboros.Consensus.Cardano.Block (EraMismatch (..))
 import qualified Ouroboros.Network.Protocol.LocalTxSubmission.Client as Net.Tx
