+++ cardano-cli/test/cardano-cli-test/Test/Cli/ITN.hs
@@ -4,7 +4,7 @@ module Test.Cli.ITN
   ( tests
   ) where
 
-import           Cardano.CLI.Shelley.Run.Key (decodeBech32)
+import           Cardano.CLI.Run.Legacy.Key (decodeBech32)
 
 import qualified Codec.Binary.Bech32 as Bech32
 import           Control.Monad (void)
